backend/requirements.txt
# ============================
# Backend Core (FastAPI stack)
# ============================

fastapi==0.115.0             # Modern async web framework for APIs and WebSockets
uvicorn[standard]==0.30.6    # ASGI server for running FastAPI apps (async + WebSocket support)

# ============================
# Configuration & Validation
# ============================

python-dotenv==1.0.1         # Load environment variables from a .env file
pydantic==2.9.2              # Data validation & schema management for FastAPI

# ============================
# Database & ORM
# ============================

sqlalchemy==2.0.34           # ORM and query builder for relational databases
psycopg2-binary==2.9.9       # PostgreSQL database adapter for SQLAlchemy

# ============================
# Cache / Message Queue
# ============================

redis==5.1.1                 # In-memory data store for caching, pub/sub, and background tasks

# ============================
# Cloud Storage / AWS
# ============================

boto3==1.35.44               # AWS SDK (S3 file storage, etc.)

# ============================
# AI / Audio Processing Stack
# ============================

# Core numeric & audio stack (pinned for Python 3.10)
numpy==1.26.4                # Compatible with PyTorch + Pyannote (avoid NumPy 2.x)
scipy==1.12.0                # Required by librosa & pyannote
librosa==0.10.1              # Audio analysis & feature extraction utilities
soundfile==0.12.1            # Audio file I/O (used internally by librosa)

# Fast Speech-to-Text
faster-whisper==1.0.3        # Optimized Whisper model (CTranslate2 backend)

# Speaker Diarization
pyannote.audio==3.1.1        # "Who spoke when" — speaker diarization model (Hugging Face pipeline)

# Deep Learning Framework
torch==2.0.1                 # PyTorch core (Windows wheel available; works with torchaudio 2.0.2)
torchaudio==2.0.2            # Audio utilities for PyTorch (needed by pyannote)
torchmetrics==1.3.1          # Metrics for PyTorch Lightning
pytorch-lightning==2.2.1     # Training utilities used internally by pyannote.audio
lightning==2.2.1             # Core Lightning package (dependency of Pyannote)
huggingface-hub==0.19.4      # HF API client (compatible with pyannote 3.1.x)

# ============================
# Optional Developer Utilities
# ============================

python-multipart==0.0.9      # Enables file uploads (if handling audio uploads via REST)
requests==2.32.3             # For external API calls or diagnostics
